+++ .vscode/settings.json
@@ -22,7 +22,9 @@
   },
   "cSpell.words": [
     "alkis",
+    "Carma",
     "flur",
+    "Geoportal",
     "landparcel"
   ]
 }
