+++ libraries/mapping/engines/cesium/measurements/src/index.ts
@@ -1,3 +1,6 @@
+// Components
+export { CesiumMeasurements } from "./lib/components/CesiumMeasurements";
+
 // Types
 export * from "./lib/types/MeasurementTypes";
